Move auth DTOs to package
