+++ src/main/java/de/tum/cit/aet/core/config/SecurityConfiguration.java
@@ -2,7 +2,7 @@ package de.tum.cit.aet.core.config;
 
 import de.tum.cit.aet.core.security.CustomJwtAuthenticationConverter;
 import de.tum.cit.aet.core.security.SpaWebFilter;
-import de.tum.cit.aet.usermanagement.dto.AuthResponseDTO;
+import de.tum.cit.aet.usermanagement.dto.auth.AuthResponseDTO;
 import de.tum.cit.aet.usermanagement.service.KeycloakAuthenticationService;
 import jakarta.servlet.http.Cookie;
 import org.springframework.context.annotation.Bean;
