--- src/main/java/de/tum/cit/aet/usermanagement/dto/AuthResponseDTO.java
@@ -1,3 +1,0 @@
-package de.tum.cit.aet.usermanagement.dto;
-
-public record AuthResponseDTO(String accessToken, String refreshToken, long expiresIn, long refreshExpiresIn) {}
