+++ src/main/java/de/tum/cit/aet/usermanagement/dto/auth/AuthResponseDTO.java
@@ -1,0 +1,4 @@
+package de.tum.cit.aet.usermanagement.dto.auth;
+
+public record AuthResponseDTO(String accessToken, String refreshToken, long expiresIn, long refreshExpiresIn) {
+}
