+++ src/main/java/de/tum/cit/aet/usermanagement/dto/auth/LoginRequestDTO.java
@@ -1,4 +1,4 @@
-package de.tum.cit.aet.usermanagement.dto;
+package de.tum.cit.aet.usermanagement.dto.auth;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
@@ -7,4 +7,5 @@ public record LoginRequestDTO(
     @Email(message = "Invalid email format") @NotBlank(message = "Email is required") String email,
 
     @NotBlank(message = "Password is required") String password
-) {}
+) {
+}
